Refactoring first step. Created communication library. And implemented the library into application-main with @LutskiYevgeniy as Navigator and @ChristoffSchaub as Driver
